+++ MonkeyDelivery/Src/Control/UI/Buttons/Continue.h
@@ -4,7 +4,7 @@
 class Continue : public Button
 {
 public:
-	Continue(int x, int y, int w, int h, Game* game);
+	Continue(int x, int y, double w, double h, Game* game);
 	~Continue() {};
 	void onCursorCollision() override;
 };
