+++ MonkeyDelivery/Src/Control/UI/Buttons/Credits.cpp
@@ -3,7 +3,7 @@
 #include "../../States/PlayingState.h"
 #include "../../States/CreditsState.h"
 
-Credits::Credits(int x, int y, int w, int h, Game* game):Button(game,w,h,x,y){
+Credits::Credits(int x, int y, double w, double h, Game* game):Button(game,w,h,x,y){
 	setTexture(button_Main_Credits);
 }
