+++ MonkeyDelivery/Src/Control/UI/Buttons/Credits.h
@@ -5,7 +5,7 @@ class Game;
 
 class Credits : public Button {
 public:
-	Credits(int x, int y,int w , int h,Game* game);
+	Credits(int x, int y, double w , double h,Game* game);
 	~Credits() {};
 	void onCursorCollision() override;
 };
