+++ MonkeyDelivery/Src/Control/UI/Buttons/Exit.h
@@ -4,7 +4,7 @@ class Game;
 
 class Exit : public Button {
 public:
-	Exit(int x, int y, int w, int h, Game* game);
+	Exit(int x, int y, double w, double h, Game* game);
 	~Exit() {};
 	void onCursorCollision() override;
 };
