+++ MonkeyDelivery/Src/Control/UI/Buttons/ExitButton.cpp
@@ -1,6 +1,6 @@
 #include "ExitButton.h"
 
-ExitButton::ExitButton(int x, int y, int w, int h, Game* game):Button(game, w, h, x, y){
+ExitButton::ExitButton(int x, int y, double w, double h, Game* game):Button(game, w, h, x, y){
 	setTexture(button_Main_ExitButton);
 }
