+++ MonkeyDelivery/Src/Control/UI/Buttons/Menu.h
@@ -6,7 +6,7 @@ class Game;
 class Menu : public Button
 {
 public:
-	Menu(int x, int y, int w, int h, Game* game);
+	Menu(int x, int y, double w, double h, Game* game);
 	~Menu() {};
 	void onCursorCollision() override;
 };
