+++ MonkeyDelivery/Src/Control/UI/Buttons/Pause.cpp
@@ -3,7 +3,7 @@
 #include "../../States/PauseState.h"
 
 
-Pause::Pause(int x, int y, int w, int h, Game* game):Button(game,w,h,x,y){
+Pause::Pause(int x, int y, double w, double h, Game* game):Button(game,w,h,x,y){
 	setTexture(button_Pause);
 }
