+++ MonkeyDelivery/Src/Control/UI/Buttons/Start.cpp
@@ -2,7 +2,7 @@
 #include "../../../Logic/Game.h"
 #include "../../States/PlayingState.h"
 
-Start::Start(int x, int y, int w, int h, Game* game):Button(game,w,h,x,y){
+Start::Start(int x, int y, double w, double h, Game* game):Button(game,w,h,x,y){
 	setTexture(button_Main_StartButton);
 }
